Bac 154 make trips persist between connections

+++ src/main/java/no/ntnu/idata2900/project/esg_module_backend/repositories/TripLogRepository.java
@@ -11,5 +11,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Group 14
  * @version v0.1.0 (2025.04.22)
  */
-public interface TripLogRepository extends JpaRepository<TripLog, Integer> {
+public interface TripLogRepository extends JpaRepository<TripLog, Long> {
 }
